Add ExtensionParserTrait and use it in `FoldersTemplateFinder` and `FileExtensionPredicate`.
Resolves #7
